Split off Code types, add #to_native()
Code types use OO to reduce duplication now.
Is in its own directoriy
to_native() converts to type.Code() - don't do that in freeze()
add check option for more rigorous code testing.
