+++ pytest/testdata/pypy_lambda.right
@@ -6,11 +6,11 @@
 # Flags:             0x00000000 (0x0)
 # First Line:        4
 # Constants:
-#    0: <xdis.code.thingy instance at 0xdeadbeef0000>
+#    0: <xdis.codetype.code2.Code2 object at 0xdeadbeef0000>
 #    1: None
 # Names:
 #    0: f
-  4:           0 LOAD_CONST                0 (<xdis.code.thingy instance at 0xdeadbeef0000>)
+  4:           0 LOAD_CONST                0 (<xdis.codetype.code2.Code2 object at 0xdeadbeef0000>)
                3 MAKE_FUNCTION             0
                6 STORE_NAME                0 (f)
                9 LOAD_CONST                1 (None)
@@ -26,7 +26,7 @@
 # First Line:        4
 # Constants:
 #    0: None
-#    1: <xdis.code.thingy instance at 0xdeadbeef0000>
+#    1: <xdis.codetype.code2.Code2 object at 0xdeadbeef0000>
 # Names:
 #    0: map
 # Varnames:
@@ -35,7 +35,7 @@
 #	right
   4:           0 BUILD_LIST                0
                3 LOOKUP_METHOD             0 (map)
-               6 LOAD_CONST                1 (<xdis.code.thingy instance at 0xdeadbeef0000>)
+               6 LOAD_CONST                1 (<xdis.codetype.code2.Code2 object at 0xdeadbeef0000>)
                9 MAKE_FUNCTION             0
               12 CALL_METHOD               1 (1 positional, 0 named)
               15 RETURN_VALUE
