+++ xdis/codetype/.gitignore
@@ -1,0 +1,1 @@
+/.mypy_cache
