+++ xdis/codetype/__init__.py
@@ -1,0 +1,22 @@
+# (C) Copyright 2017-2020 by Rocky Bernstein
+#
+#  This program is free software; you can redistribute it and/or
+#  modify it under the terms of the GNU General Public License
+#  as published by the Free Software Foundation; either version 2
+#  of the License, or (at your option) any later version.
+#
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with this program; if not, write to the Free Software
+#  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+
+__docformat__ = "restructuredtext"
+from xdis.codetype.base import *
+from xdis.codetype.code14 import *
+from xdis.codetype.code2 import *
+from xdis.codetype.code3 import *
+from xdis.codetype.code38 import *
