+++ xdis/marsh.py
@@ -30,7 +30,7 @@ there). Details of the format may change between Python versions.
 import types, struct
 
 from xdis.version_info import PYTHON_VERSION, PYTHON3
-from xdis.code import Code2, Code2Compat, Code3, Code3Compat
+from xdis.codetype import Code2, Code2Compat, Code3, Code3Compat
 
 try:
     intern
